[fix] 월별 record 조회API사용시 요청URL 오류 수정

+++ front/app/src/main/java/com/ssafy/oringe/api/record/Record.java
@@ -17,13 +17,11 @@ import lombok.ToString;
 @ToString
 public class Record {
 
-    private Long recordId;
+    private int recordId;
 
     private Boolean recordSuccess;
 
     private LocalDate recordDate;
-
-
 
 //    private List<String> recordTemplates;
 }
